+++ lang/syn/src/codegen/accounts/to_account_infos.rs
@@ -1,11 +1,16 @@
-use crate::codegen::accounts::generics;
+use crate::codegen::accounts::{generics, ParsedGenerics};
 use crate::{AccountField, AccountsStruct};
 use quote::quote;
 
 // Generates the `ToAccountInfos` trait implementation.
 pub fn generate(accs: &AccountsStruct) -> proc_macro2::TokenStream {
     let name = &accs.ident;
-    let (combined_generics, trait_generics, strct_generics) = generics(accs);
+    let ParsedGenerics {
+        combined_generics,
+        trait_generics,
+        struct_generics,
+        where_clause,
+    } = generics(accs);
 
     let to_acc_infos: Vec<proc_macro2::TokenStream> = accs
         .fields
@@ -21,7 +26,8 @@ pub fn generate(accs: &AccountsStruct) -> proc_macro2::TokenStream {
         })
         .collect();
     quote! {
-        impl#combined_generics anchor_lang::ToAccountInfos#trait_generics for #name#strct_generics {
+        #[automatically_derived]
+        impl<#combined_generics> anchor_lang::ToAccountInfos<#trait_generics> for #name <#struct_generics> #where_clause{
             fn to_account_infos(&self) -> Vec<anchor_lang::solana_program::account_info::AccountInfo<'info>> {
                 let mut account_infos = vec![];